# src/swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Med Record | API Documentation",
    "description": "Documentation for the Med Record API, an example of a medical record system.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "description": "Local server",
      "url": "http://localhost:3001"
    },
    {
      "description": "Production server",
      "url": "https://med-record.goncalvesmarques.com"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/doctors/{id}": {
      "get": {
        "summary": "Get doctor by ID",
        "tags": [
          "Doctors"
        ],
        "operationId": "GetDoctorByIdUseCase",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the doctor",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorNotFound"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update doctor by ID",
        "tags": [
          "Doctors"
        ],
        "operationId": "UpdateDoctorUseCase",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the doctor",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorNotFound"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete doctor by ID",
        "tags": [
          "Doctors"
        ],
        "operationId": "DeleteDoctorByIdUseCase",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the doctor",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Doctor deleted successfully."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Doctor not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorNotFound"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "InternalServerError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Internal server error"
          }
        }
      },
...
      "UnauthorizedError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "UNAUTHORIZED"
          },
          "message": {
            "type": "string",
            "example": "Missing token"
          },
          "service": {
            "type": "string",
            "example": "AuthMiddleware"
          }
        }
      },
...
      "DoctorResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Dr. House"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "dr.house@mail.com"
          }
        }
      },
      "DoctorNotFound": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "code": {
            "type": "string",
            "example": "DOCTOR_NOT_FOUND_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Doctor not found"
          },
          "service": {
            "type": "string",
            "example": "DoctorController"
          }
        }
      },
      "DoctorUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Dr. House MD"
          }
        }
      },
...
    }
  }
}